+++ nisapi/clean/udsf_9v7b.py
@@ -1,6 +1,7 @@
 import polars as pl
 import uuid
-import nisapi.clean
+import calendar
+from nisapi.clean import admin1_values, drop_suppressed_rows
 
 
 def _clean_geography_expr(type_: pl.Expr, value: pl.Expr) -> pl.Expr:
@@ -10,8 +11,7 @@ def _clean_geography_expr(type_: pl.Expr, value: pl.Expr) -> pl.Expr:
         .when(type_ == pl.lit("HHS Regional Estimates"))
         .then(pl.lit("region"))
         .when(
-            (type_ == pl.lit("Jurisdictional Estimates"))
-            & (value.is_in(nisapi.clean.admin1_values))
+            (type_ == pl.lit("Jurisdictional Estimates")) & (value.is_in(admin1_values))
         )
         .then(pl.lit("admin1"))
         .when(
@@ -54,6 +54,56 @@ def clean_region(x: pl.Expr) -> pl.Expr:
     return x.str.extract("^(Region \\d+): ", 1)
 
 
+def parse_coninf_95(df: pl.DataFrame) -> pl.DataFrame:
+    df_split = df.with_columns(
+        pl.col("coninf_95")
+        .str.split_exact(" - ", 1)
+        .struct.rename_fields(["lci", "uci"])
+    )
+
+    return df_split.unnest("coninf_95")
+
+
+def month_name_to_number(x: pl.Expr) -> pl.Expr:
+    # note that we need to do this union because "May" is both a full name and an abbreviation,
+    # and replace_strict wants unique old values. Note also that range(13) includes 0, and
+    # month_name[0] is ""
+    mapping = dict(zip(calendar.month_name, range(13))) | dict(
+        zip(calendar.month_abbr, range(13))
+    )
+    return x.replace_strict(mapping)
+
+
+def _parse_time_period_expr(time_year: pl.Expr, time_period: pl.Expr) -> pl.Expr:
+    year = time_year.cast(pl.Int32)
+    period_split = time_period.str.extract_groups(
+        r"^(\w+)\s+(\w+)\s+-\s+(\w+)\s+(\w+)\s*$"
+    ).struct.rename_fields(["month1", "day1", "month2", "day2"])
+
+    month1 = period_split.struct["month1"].pipe(month_name_to_number)
+    day1 = period_split.struct["day1"].str.strip_chars().cast(pl.Int32)
+    month2 = period_split.struct["month2"].pipe(month_name_to_number)
+    day2 = period_split.struct["day2"].str.strip_chars().cast(pl.Int32)
+
+    date1 = pl.date(year, month1, day1)
+    date2 = pl.date(year, month2, day2)
+
+    return pl.struct(start=date1, end=date2)
+
+
+def parse_time_period(df: pl.DataFrame) -> pl.DataFrame:
+    column_name = str(uuid.uuid1())
+    return (
+        df.with_columns(
+            _parse_time_period_expr(pl.col("time_year"), pl.col("time_period")).alias(
+                column_name
+            )
+        )
+        .unnest(column_name)
+        .drop(["time_year", "time_period"])
+    )
+
+
 def clean(df: pl.LazyFrame) -> pl.LazyFrame:
     return (
         df.rename(
@@ -66,10 +116,10 @@ def clean(df: pl.LazyFrame) -> pl.LazyFrame:
                 "indicator_category": "indicator_value",
             }
         )
-        .pipe(nisapi.clean.drop_suppressed_rows)
+        .pipe(drop_suppressed_rows)
         .drop("sample_size")
-        .pipe(nisapi.clean.parse_coninf_95)
-        .pipe(nisapi.clean.parse_time_period)
+        .pipe(parse_coninf_95)
+        .pipe(parse_time_period)
         .with_columns(
             pl.col("time_type").replace_strict({"Monthly": "month", "Weekly": "week"})
         )
